+++ ios/ios/src/Users/CMUser.h
@@ -160,6 +160,22 @@ typedef void (^CMUserFetchCallback)(NSArray *users, NSDictionary *errors);
 - (void)loginWithCallback:(CMUserOperationCallback)callback;
 
 /**
+ * Login with social networking sites through Singly.  Calls a UIWebView for authenticaion
+ *
+ * Possible result codes:
+ * - <tt>CMUserAccountLoginSucceeded</tt>
+ * - <tt>CMUserAccountLoginFailedIncorrectCredentials</tt>
+ * - <tt>CMUserAccountOperationFailedUnknownAccount</tt>
+ *
+ * @param service The social service to be logged into
+ * @param viewController the current view controller in use when this method is called
+ * @param params Can be nil. These will be added to the end of the call for authentication, and is a good place to put the scope. This is simply appended to the back of the URL if present, so it is up to the caller to ensure the parameters are of the correct format.
+ * @param callback The block that will be called on completion of the operation.
+ * @see https://cloudmine.me/docs/api#users_social
+ */
+- (void)loginWithSocialNetwork:(NSString *)service viewController:(UIViewController *)viewController params:(NSString *)params callback:(CMUserOperationCallback)callback;
+
+/**
  * Asynchronously logout the user and clear their session and session token. On completion, the <tt>callback</tt> block will be called with
  * the result of the operation and any messages returned by the server contained in an array. See the CloudMine
  * documentation online for the possible contents of this array.
@@ -268,22 +284,6 @@ typedef void (^CMUserFetchCallback)(NSArray *users, NSDictionary *errors);
  * @param callback The block that will be called on completion of the operation.
  */
 - (void)save:(CMUserOperationCallback)callback;
-
-
-/**
- * Login with social networking sites through Singly.  Calls a UIWebView for authenticaion
- *
- * Possible result codes:
- * - <tt>CMUserAccountLoginSucceeded</tt>
- * - <tt>CMUserAccountLoginFailedIncorrectCredentials</tt>
- * - <tt>CMUserAccountOperationFailedUnknownAccount</tt>
- *
- * @param callback The block that will be called on completion of the operation.
- * @param service The social service to be logged into
- * @param viewController the current view controller in use when this method is called
- * @see https://cloudmine.me/docs/api#users_social
- */
-- (void)loginWithSocial:(NSString *)service andViewController:(UIViewController *)viewController callback:(CMUserOperationCallback)callback;
 
 /**
  * Asynchronously fetch all the users of this app. This will download the profiles of all the users of your app, and is useful for displaying